Rename transaction selectors file
