+++ ui/components/app/transaction-list/transaction-list.component.js
@@ -24,7 +24,7 @@ import {
   nonceSortedCompletedTransactionsSelectorAllChains,
   nonceSortedPendingTransactionsSelector,
   nonceSortedPendingTransactionsSelectorAllChains,
-} from '../../../selectors/transactions';
+} from '../../../selectors/transactions-legacy';
 import { getCurrentChainId } from '../../../../shared/modules/selectors/networks';
 import {
   getCurrentNetwork,
