+++ ui/selectors/selectors.js
@@ -136,7 +136,7 @@ import {
   getAllUnapprovedTransactions,
   getCurrentNetworkTransactions,
   getUnapprovedTransactions,
-} from './transactions';
+} from './transactions-legacy';
 // eslint-disable-next-line import/order
 import { getSelectedInternalAccount, getInternalAccounts } from './accounts';
 import {
